# docs/openapi.yaml
openapi: 3.0.3
info:
  title: KubeWise API
  description: Kubernetes intelligent operations agent system REST API
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /api/v1/chat/confirm:
    post:
      summary: Confirm operation step
      description: |
        Respond to a `confirm_request` SSE event. The agent is blocked until this endpoint is called.
      operationId: chatConfirm
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmRequest"
      responses:
        "200":
          description: Confirmation accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: confirm_id not found or already responded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Agent no longer waiting for confirmation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ConfirmRequest:
      type: object
      required:
        - confirm_id
        - confirmed
      properties:
        confirm_id:
          type: string
          description: ID from the confirm_request SSE event
        confirmed:
          type: boolean
          description: true to execute the step, false to skip
        correction:
          type: string
          description: Optional correction instructions (when confirmed is false)
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        detail:
          type: string